backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.scripts]
generate = "app.engine.generate:generate_datasource"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
aiostream = "^0.5.2"
llama-index = "0.10.58"
cachetools = "^5.3.3"
spacy = "^3.7.5"
en_core_web_sm ="^3.7.1"
python-frontmatter = "^1.1.0"
voyageai = "^0.2.3"
scipy = "^1.14.1"
langfuse = "^2.52.1"
arize-phoenix = "^5.5.2"
llama-index-core = ">=0.10.58"
openinference-instrumentation-llama-index = ">=2"
opentelemetry-proto = ">=1.12.0"
arize-phoenix-otel = "^0.6.1"
pytector = "^0.2.0"
requests = "^2.32.5"
langdetect = "^1.0.9"
psutil = "^5.9.8"
pyarrow = "^14.0.1"
pandas = "^2.1.4"
boto3 = "^1.34.34"
apscheduler = "^3.10.4"
httpx = "^0.27.0"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.gunicorn]
version = "^21.2.0"

[tool.poetry.dependencies.llama-index-vector-stores-pinecone]
version = "^0.1.3"

[tool.poetry.dependencies.docx2txt]
version = "^0.8"

[tool.poetry.dependencies.llama-index-agent-openai]
version = "0.2.6"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.5"
pytest = "^8.4.2"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"